Switched from UnitTest to pytest, revised file structure, updated README, added .gitignore, CONTRIBUTORS.md, and requirements.txt
