+++ CONTRIBUTORS.md
@@ -1,0 +1,9 @@
+# Contributors
+
+We would like to thank the following people for their contributions to this project:
+
+- **Aya Swazuz** - Project Creator
+- **Glenn Lehman** - Changed project structure and testing to support pyTest, revised README.md, added CONTRIBUTING.md, and added CONTRIBUTORS.md
+
+
+If you have contributed to this project and your name is not listed here, please feel free to add it by submitting a pull request.
